created APIs to send data to buffer... error copiling when calling functions in eth.h in queue.cpp... main in eth.cpp for now
