+++ us/bin/network/include.mk
@@ -1,4 +1,4 @@
-NETWORK_SRCS=$(addprefix us/bin/network/,queue.cpp)
+NETWORK_SRCS=$(addprefix us/bin/network/,eth.cpp)
 NETWORK_OBJS=$(addprefix $(BUILDDIR)/,$(NETWORK_SRCS:.cpp=.o))
 
 #NETWORK_LIBS=-Wl,--whole-archive -lbacktrace -Wl,--no-whole-archive
